Fix naming and structure of service files
