--- MovieQuiz/Services/QuestionFactoryDelegate.swift
@@ -1,14 +1,0 @@
-//
-//  QuestionFactoryDelegate.swift
-//  MovieQuiz
-//
-//  Created by Andrei Chenchik on 11/8/22.
-//
-
-import Foundation
-
-protocol QuestionFactoryDelegate: AnyObject {
-    func didReceiveNextQuestion(question: QuizQuestion?)
-    func didLoadDataFromServer()
-    func didFailToLoadData(with error: Error)
-}
