--- MovieQuiz/Services/QuestionFactoryProtocol.swift
@@ -1,13 +1,0 @@
-//
-//  QuestionFactoryProtocol.swift
-//  MovieQuiz
-//
-//  Created by Andrei Chenchik on 11/8/22.
-//
-
-import Foundation
-
-protocol QuestionFactoryProtocol {
-    func loadData()
-    func requestNextQuestion()
-}
